Add transpose & repeat kernel for c++ for opencl (#258)
* add transpose kernel for c++ for opencl

* add repeat kernel for c++ for opencl

* disable utl::static_vector copy constructor on opencl

* fix transpose & repeat view for c++ for opencl

* update c++ for opencl tests

* update repeat tests

* update opencl copy_buffer to handle numeric type

* init size_type specification for index functions

* add test data for repeat and transpose indexing

* fix repeat and transpose opencl kernels

* add cast index function

* update repeat and transpose view for opencl

* update opencl tester to cast index array

* update tests

* fix opencl index tests

+++ include/nmtools/array/index/product.hpp
@@ -36,7 +36,7 @@ namespace nmtools::index
                     ret = p;
                 });
             } else {
-                for (size_t i=0; i<len(shape); i++) {
+                for (size_t i=0; i<(size_t)len(shape); i++) {
                     auto p = ret * at(shape,i);
                     ret = p;
                 }
